move shutdown to its own module
