+++ TestBoardFirmware/phantom_drivers/common/source/validation.c
@@ -11,8 +11,6 @@
 #define LOW_BOUND 0
 #define UP_BOUND 1
 
-#define SHUTDOWN_TIMEOUT_PERIOD 5000
-
 //Lookup tables for expected results from the VCU
 
 //TODO: Fill out ranges for expected results ie { MIN, MAX } -> { 10, 20 } =  10 <= expected result <= 20
@@ -55,14 +53,6 @@ static const unsigned int BSE_RESULTS_LUT[][2]  = {  { 0, 20 }, //NORMAL_BSE_OFF
 
 //static const uint8_t[] GPIO_RESULTS_LUT = { /* RUNNING */ 0 , /* TS_OFF */0, 0 /* !RUNNING */, /* ZERO THROTTLE */ 0}; //wait for CAN driver...
 
-
-/* Shutdown Variables */
-
-static shutdown_fail;
-
-static shutdown_signal;
-
-static shutdown_expected;
 
 static bool validateAPPS(uint8_t, unsigned int);
 
@@ -113,101 +103,3 @@ static bool validateBSE(uint8_t testcase, unsigned int inverter_sig){
     //check if received value is within range of result
     return ( (BSE_RESULTS_LUT[testcase][LOW_BOUND] <= inverter_sig) && (inverter_sig <= BSE_RESULTS_LUT[testcase][UP_BOUND]) );
 }
-
-
-/* Shutdown Interrupts */
-
-//Halcogen Setup
-//1.Enable HET1 Driver
-//2.Choose Edge Interrupt to enable and output on specific pin
-//3.Select VIM Channel for either Low/High Priority (same priority as edge interrupt)
-//4.Initialize VIM,GIO, and enableIRQ in the code
-//HET1 interrupt set at both edges; set to high priority
-
-void initializeShutdownInterrupt(){
-
-    hetInit();
-
-    vimInit();
-
-    //Enable Interrupts
-    edgeEnableNotification(VCU_FLT_REG, VCU_FLT_EDGE);
-
-    _enable_IRQ();
-
-}
-
-static void readShutdownSignal(){
-
-    shutdown_signal = gioGetBit( VCU_FLT_PORT, VCU_FLT_PIN );
-}
-
-
-void shutdown_timeout_callback(int ID){
-
-    shutdown_fail = true;
-
-    UARTprintf("SHUTDOWN SIGNAL HAS TIMED OUT\r\n");
-
-    stopAllTimers();
-
-}
-
-void shutdown_callback(){
-
-    if( shutdown_expected == getShutdownSignal() ){
-
-        initializeVCU();
-
-
-    }else{
-
-        stopAllTimers();
-
-        shutdown_fail = true;
-
-    }
-
-}
-
-
-uint8_t getShutdownSignal(){
-
-    return shutdown_signal;
-}
-
-void setShutdownOccurence(bool expected_result){
-
-    setTimerID(VALIDATION, 0)
-
-
-    setTimerCallback(VALIDATION, shutdown_timeout_callback);
-
-    startTimer(VALIDATION, SHUTDOWN_TIMEOUT_PERIOD);
-
-    shutdown_expected = expected_result;
-}
-
-bool getShutdownResult(){
-
-    return shutdown_fail;
-}
-
-void edgeNotification(hetBASE_t * hetREG,uint32 edge)
-{
-/*  enter user code between the USER CODE BEGIN and USER CODE END. */
-/* USER CODE BEGIN (37) */
-
-    if(hetREG == VCU_FLT_REG && edge == VCU_FLT_EDGE){
-
-        readShutdownSignal();
-
-        shutdown_callback();
-    }
-
-
-
-/* USER CODE END */
-}
-
-
